+++ test/integration_tests/integrationtest_ros_scanner_node.cpp
@@ -15,7 +15,6 @@
 
 #include <memory>
 #include <chrono>
-#include <thread>
 #include <future>
 
 #include <gtest/gtest.h>
@@ -46,7 +45,6 @@ using namespace psen_scan_v2_standalone;
 using namespace psen_scan_v2_test;
 
 using namespace ::testing;
-using namespace psen_scan_v2_standalone::configuration;
 using namespace psen_scan_v2_standalone_test;
 
 namespace psen_scan_v2
@@ -134,8 +132,7 @@ protected:
 
 TEST_F(RosScannerNodeTests, shouldStartAndStopSuccessfullyIfScannerRespondsToRequests)
 {
-  ROSScannerNodeT<ScannerMock> ros_scanner_node(
-      nh_priv_, "scan", "scanner", configuration::DEFAULT_X_AXIS_ROTATION, scanner_config_);
+  ROSScannerNodeT<ScannerMock> ros_scanner_node(nh_priv_, "scan", "scanner", 1.0 /*x_axis_rotation*/, scanner_config_);
 
   std::promise<void> hw_finished_request;
   hw_finished_request.set_value();
@@ -157,10 +154,41 @@ TEST_F(RosScannerNodeTests, shouldStartAndStopSuccessfullyIfScannerRespondsToReq
   EXPECT_FUTURE_IS_READY(loop, LOOP_END_TIMEOUT);
 }
 
+TEST_F(RosScannerNodeTests, shouldThrowExceptionSetInScannerStartFuture)
+{
+  ROSScannerNodeT<ScannerMock> ros_scanner_node(nh_priv_, "scan", "scanner", 1.0 /*x_axis_rotation*/, scanner_config_);
+
+  std::promise<void> hw_finished_request;
+  const std::string error_msg = "error msg for testing";
+  hw_finished_request.set_exception(std::make_exception_ptr(std::runtime_error(error_msg)));
+
+  ON_CALL(ros_scanner_node.scanner_, start()).WillByDefault(Return_Future(hw_finished_request));
+  EXPECT_THROW_AND_WHAT(ros_scanner_node.run(), std::runtime_error, error_msg.c_str());
+}
+
+TEST_F(RosScannerNodeTests, shouldThrowDelayedExceptionSetInScannerStartFuture)
+{
+  ROSScannerNodeT<ScannerMock> ros_scanner_node(nh_priv_, "scan", "scanner", 1.0 /*x_axis_rotation*/, scanner_config_);
+
+  util::Barrier start_barrier;
+  std::promise<void> hw_finished_request;
+  ON_CALL(ros_scanner_node.scanner_, start())
+      .WillByDefault(DoAll(OpenBarrier(&start_barrier), Return_Future(hw_finished_request)));
+
+  std::future<void> loop = std::async(std::launch::async, [&ros_scanner_node]() { ros_scanner_node.run(); });
+  start_barrier.waitTillRelease(DEFAULT_TIMEOUT);
+  ros::Duration(1.0).sleep();
+
+  const std::string error_msg = "error msg for testing";
+  hw_finished_request.set_exception(std::make_exception_ptr(std::runtime_error(error_msg)));
+
+  EXPECT_FUTURE_IS_READY(loop, LOOP_END_TIMEOUT);
+  EXPECT_THROW_AND_WHAT(loop.get(), std::runtime_error, error_msg.c_str());
+}
+
 TEST_F(RosScannerNodeTests, shouldProvideScanTopic)
 {
-  ROSScannerNodeT<ScannerMock> ros_scanner_node(
-      nh_priv_, "scan", "scanner", configuration::DEFAULT_X_AXIS_ROTATION, scanner_config_);
+  ROSScannerNodeT<ScannerMock> ros_scanner_node(nh_priv_, "scan", "scanner", 1.0 /*x_axis_rotation*/, scanner_config_);
 
   util::Barrier start_barrier;
   util::Barrier stop_barrier;
@@ -178,8 +206,7 @@ TEST_F(RosScannerNodeTests, shouldProvideScanTopic)
 
 TEST_F(RosScannerNodeTests, shouldProvideActiveZonesetTopic)
 {
-  ROSScannerNodeT<ScannerMock> ros_scanner_node(
-      nh_priv_, "scan", "scanner", configuration::DEFAULT_X_AXIS_ROTATION, scanner_config_);
+  ROSScannerNodeT<ScannerMock> ros_scanner_node(nh_priv_, "scan", "scanner", 1.0 /*x_axis_rotation*/, scanner_config_);
 
   util::Barrier start_barrier;
   util::Barrier stop_barrier;
@@ -197,8 +224,7 @@ TEST_F(RosScannerNodeTests, shouldProvideActiveZonesetTopic)
 
 TEST_F(RosScannerNodeTests, shouldPublishScansWhenLaserScanCallbackIsInvoked)
 {
-  ROSScannerNodeT<ScannerMock> ros_scanner_node(
-      nh_priv_, "scan", "scanner", configuration::DEFAULT_X_AXIS_ROTATION, scanner_config_);
+  ROSScannerNodeT<ScannerMock> ros_scanner_node(nh_priv_, "scan", "scanner", 1.0 /*x_axis_rotation*/, scanner_config_);
 
   util::Barrier scan_topic_barrier;
   SubscriberMock subscriber(nh_priv_);
@@ -226,8 +252,7 @@ TEST_F(RosScannerNodeTests, shouldPublishScansWhenLaserScanCallbackIsInvoked)
 
 TEST_F(RosScannerNodeTests, shouldPublishActiveZonesetWhenLaserScanCallbackIsInvoked)
 {
-  ROSScannerNodeT<ScannerMock> ros_scanner_node(
-      nh_priv_, "scan", "scanner", configuration::DEFAULT_X_AXIS_ROTATION, scanner_config_);
+  ROSScannerNodeT<ScannerMock> ros_scanner_node(nh_priv_, "scan", "scanner", 1.0 /*x_axis_rotation*/, scanner_config_);
 
   const uint8_t first_zone{ 2 };
   const uint8_t second_zone{ 4 };
@@ -258,8 +283,7 @@ TEST_F(RosScannerNodeTests, shouldPublishActiveZonesetWhenLaserScanCallbackIsInv
 
 TEST_F(RosScannerNodeTests, shouldReceiveLatchedActiveZonesetMsg)
 {
-  ROSScannerNodeT<ScannerMock> ros_scanner_node(
-      nh_priv_, "scan", "scanner", configuration::DEFAULT_X_AXIS_ROTATION, scanner_config_);
+  ROSScannerNodeT<ScannerMock> ros_scanner_node(nh_priv_, "scan", "scanner", 1.0 /*x_axis_rotation*/, scanner_config_);
   util::Barrier start_barrier;
   util::Barrier stop_barrier;
   EXPECT_SUCCESSFULL_START_AND_OPEN_BARRIER_ON_CALL(ros_scanner_node.scanner_, &start_barrier);
@@ -314,8 +338,7 @@ TEST_F(RosScannerNodeTests, shouldPublishScanEqualToConversionOfSuppliedLaserSca
 
 TEST_F(RosScannerNodeTests, shouldWaitWhenStopRequestResponseIsMissing)
 {
-  ROSScannerNodeT<ScannerMock> ros_scanner_node(
-      nh_priv_, "scan", "scanner", configuration::DEFAULT_X_AXIS_ROTATION, scanner_config_);
+  ROSScannerNodeT<ScannerMock> ros_scanner_node(nh_priv_, "scan", "scanner", 1.0 /*x_axis_rotation*/, scanner_config_);
 
   util::Barrier start_barrier;
   EXPECT_SUCCESSFULL_START_AND_OPEN_BARRIER_ON_CALL(ros_scanner_node.scanner_, &start_barrier);